requirements.txt
streamlit==1.25.0
opencv-python==4.8.0.74
Pillow==9.5.0